Fallback to usable when physical mem not determined
Sometimes the physical memory cannot be determined without root privileges
because either the syslog doesn't exist or the user does not have the
privileges to read it. This can happen in containerized or reduced-privilege
environments. In these cases, fall back to using usable amount of memory, which
is always readable from /proc/meminfo.

Closes issue #2
Partly addresses issue #1